+++ include/command/WithdrawFundsCommand.h
@@ -4,8 +4,8 @@
 
 class WithdrawFundsCommand: public ICommand {
 public:
-    WithdrawFundsCommand(int user_id, int quantity);
+    WithdrawFundsCommand(int user_id, double quantity);
 private:
     int user_id_;
-    int quantity_;
+    double quantity_;
 };
