+++ source/command/DepositFundsCommand.cpp
@@ -1,4 +1,4 @@
 #include "../../include/command/DepositFundsCommand.h"
 
-DepositFundsCommand::DepositFundsCommand(int user_id, int quantity)
+DepositFundsCommand::DepositFundsCommand(int user_id, double quantity)
         : user_id_(user_id), quantity_(quantity) {}
